segurança contra tentar excluir sem estar logado

+++ crud/deletaratividade.php
@@ -3,12 +3,18 @@
 
     $id = $_GET['id'];
 
-    $sql = "DELETE FROM atividades WHERE id = $id";
+    session_start();
+    if(isset($_SESSION["user"]) && $_SESSION["user"] != 0){
 
-    if ($conn->query($sql) === TRUE) {
-        header("Location: http://localhost/HugAll-main/perfil.php");
-    }
-    else {
+        $sql = "DELETE FROM atividades WHERE id = $id";
+
+        if ($conn->query($sql) === TRUE) {
+            header("Location: http://localhost/HugAll-main/perfil.php");
+        }
+        else {
+            header("Location: http://localhost/HugAll-main/index.php");
+        }
+    } else {
         header("Location: http://localhost/HugAll-main/index.php");
     }
